Перенос сотрудников в проект доменных моделей
