+++ WebStore.Domain/Entities/Employee.cs
@@ -1,0 +1,17 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace WebStore.Domain.Entities;
+
+public class Employee
+{
+	public int Id { get; set; }
+	public string LastName { get; set; }
+	public string FirstName { get; set; }
+	public string Patronymic { get; set; }
+	public int Age { get; set; }
+
+}
