+++ WebStore/Views/Employees/Details.cshtml
@@ -1,4 +1,4 @@
-﻿@model WebStore.Models.Employee
+﻿@model 			Employee
 
 @{
     ViewData["Title"] = "Информация о сотруднике id: " + Model.Id;
