+++ WebStore/Views/Employees/Index.cshtml
@@ -1,4 +1,4 @@
-﻿@model IEnumerable<WebStore.Models.Employee>
+﻿@model IEnumerable<Employee>
 
 @{
     ViewData["Title"] = "Список сотрудников";
